+++ CHANGES.md
@@ -17,7 +17,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 % Security   Security in case of vulnerabilities.
 
 See also the
-[unreleased changes](https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.6.2...branch%2Fdefault).
+[unreleased changes](https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.6.3...branch%2Fdefault).
+
+## [0.6.3] (2024-04-22)
+
+- New functions {func}`fluiddyn.clusters.Cluster.get_commands_setting_env` and
+  {func}`fluiddyn.clusters.Cluster.get_commands_activating_lauching_python`.
 
 ## [0.6.2] (2024-04-16)
 
@@ -241,3 +246,4 @@ Improvements and refactoring {mod}`fluiddyn.util.serieofarrays` with much better
 [0.6.0]: https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.5.4...0.6.0
 [0.6.1]: https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.6.0...0.6.1
 [0.6.2]: https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.6.1...0.6.2
+[0.6.3]: https://foss.heptapod.net/fluiddyn/fluiddyn/-/compare/0.6.2...0.6.3